# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "O-Ushers API",
    "description": "The **O-Ushers** platform connects event organizers with professional ushers (talents).\n\n## Roles\n| Role | Description |\n|------|-------------|\n| **Usher** | A talent who browses events, applies, and manages their profile |\n| **Organizer** | Posts events, manages applicants, marks attendance & reviews |\n| **Admin** | Platform-level management: users, events, verification |\n\n## Authentication\nAll protected endpoints require a **Bearer JWT** token obtained from `POST /auth/login`.\n\n## Pagination\nPaginated endpoints accept `page` and `size` query parameters and return:\n```json\n{ \"data\": [], \"currentPage\": 1, \"totalPages\": 5, \"totalCount\": 42 }\n```\n",
    "version": "1.0.0",
    "contact": {
      "email": "ooushers@gmail.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    },
    {
      "url": "https://o-ushers.vercel.app",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/usher/{id}": {
      "get": {
        "tags": [
          "Ushers"
        ],
        "summary": "Get own profile",
        "description": "Returns the authenticated usher's own full profile.\n`id` in the path must match the authenticated user's ID.\n",
        "operationId": "getUsherProfile",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The usher's own UUID (must match auth token)."
          }
        ],
        "responses": {
          "200": {
            "description": "Own usher profile",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/User"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/usher/profile/{id}": {
      "get": {
        "tags": [
          "Ushers"
        ],
        "summary": "View any usher's public profile",
        "description": "Returns a public, read-only view of a usher profile including event history\nand reviews. Accessible by **usher**, **organizer**, and **admin** roles.\n",
        "operationId": "getUsherProfileById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "UUID of the usher to view."
          }
        ],
        "responses": {
          "200": {
            "description": "Public usher profile with event history and reviews",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "user": {
                              "$ref": "#/components/schemas/User"
                            },
                            "eventHistory": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/EventHistoryItem"
                              }
                            },
                            "reviews": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/Review"
                              }
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/usher/events/browse": {
      "get": {
        "tags": [
          "Ushers"
        ],
        "summary": "Browse available events",
        "description": "Returns a paginated list of currently **open** events whose application\ndeadline has not passed. Filterable by `category` and `city`.\n",
        "operationId": "browseEvents",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "wedding",
                "corporate",
                "club",
                "festival",
                "private_party",
                "conference",
                "exhibition",
                "sport_event"
              ]
            },
            "description": "Filter by event category."
          },
          {
            "name": "city",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Filter by city name (case-insensitive partial match)."
          },
          {
            "$ref": "#/components/parameters/PageParam"
          },
          {
            "$ref": "#/components/parameters/SizeParam"
          },
          {
            "$ref": "#/components/parameters/SortParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of open events",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedEvents"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/usher/events/apply": {
      "post": {
        "tags": [
          "Ushers"
        ],
        "summary": "Apply to an event",
        "description": "Submits an application to an open event. Business rules enforced:\n- **BR-01**: Cannot apply twice to the same event.\n- **BR-02**: Application deadline must not have passed.\n",
        "operationId": "applyToEvent",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "eventId"
                ],
                "properties": {
                  "eventId": {
                    "type": "string",
                    "format": "uuid",
                    "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Application submitted",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Application"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/usher/applications/my": {
      "get": {
        "tags": [
          "Ushers"
        ],
        "summary": "Track my applications",
        "description": "Returns the usher's own applications with enriched event data.\nOptionally filter by `upcoming` or `past` events.\n",
        "operationId": "getMyApplications",
        "parameters": [
          {
            "name": "filter",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "upcoming",
                "past"
              ]
            },
            "description": "Filter applications by event date."
          },
          {
            "$ref": "#/components/parameters/PageParam"
          },
          {
            "$ref": "#/components/parameters/SizeParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated applications list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedApplications"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/usher/events/history": {
      "get": {
        "tags": [
          "Ushers"
        ],
        "summary": "Get event history",
        "description": "Returns all past events where the usher was **accepted** or **excused**,\nwith attendance status and review rating for each event.\n",
        "operationId": "getMyEventHistory",
        "responses": {
          "200": {
            "description": "Event history list",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/EventHistoryItem"
                          }
                        },
                        "count": {
                          "type": "integer",
                          "example": 7
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/usher/applications/{applicationId}/excuse": {
      "patch": {
        "tags": [
          "Ushers"
        ],
        "summary": "Excuse from an accepted event",
        "description": "Marks an accepted application as **excused**. Business rules:\n- **BR-03**: Can only excuse from `accepted` applications.\n- **BR-04**: Excusing within 3 days of deadline = late excuse penalty (+1 counter).\n- **BR-08**: Late excuse resets the consecutive-good-events streak.\n",
        "operationId": "excuseFromEvent",
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "UUID of the application to excuse."
          }
        ],
        "responses": {
          "200": {
            "description": "Excused successfully (with or without late penalty)",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "applicationStatus": {
                              "type": "string",
                              "example": "excused"
                            },
                            "isLateExcuse": {
                              "type": "boolean"
                            },
                            "lateExcuseCount": {
                              "type": "integer"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/usher/refer": {
      "post": {
        "tags": [
          "Ushers"
        ],
        "summary": "Refer a talent to an event",
        "description": "Refers another usher to an open event. Business rules:\n- **BR-13**: Cannot refer a talent already accepted or rejected for the event.\n- Cannot refer yourself.\n- Creates an application (pending) for the referred talent if none exists.\n",
        "operationId": "referTalent",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "eventId",
                  "referredTalentId"
                ],
                "properties": {
                  "eventId": {
                    "type": "string",
                    "format": "uuid",
                    "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                  },
                  "referredTalentId": {
                    "type": "string",
                    "format": "uuid",
                    "example": "6ba45f12-1234-4562-c3ab-9f963f44bfa4"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Talent referred successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Referral"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from `POST /auth/login`. Pass it in the\n`Authorization` header as `Bearer <token>`.\n"
      }
    },
...
    "responses": {
...
      "Unauthorized": {
        "description": "Missing or invalid Bearer token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "Authentication required"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Authenticated but not authorized for this action",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "Insufficient permissions"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "User not found"
            }
          }
        }
      },
...
    },
    "schemas": {
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Operation completed successfully"
          }
        }
      },
...
      "User": {
        "type": "object",
        "description": "User model. Sensitive fields (password, otp, otpExpiry, otpAttempts,\nlastOtpRequest, otpVerified) are never returned in responses.\n",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "6ba45f12-1234-4562-c3ab-9f963f44bfa4"
          },
          "fullName": {
            "type": "string",
            "example": "Ahmed Ali"
          },
          "userName": {
            "type": "string",
            "example": "ahmed_ali"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "ahmed@example.com"
          },
          "mobileNumber": {
            "type": "string",
            "example": "+201234567890"
          },
          "city": {
            "type": "string",
            "example": "Cairo"
          },
          "role": {
            "type": "string",
            "enum": [
              "usher",
              "organizer",
              "admin"
            ]
          },
          "experience": {
            "type": "integer",
            "example": 3
          },
          "rate": {
            "type": "number",
            "format": "float",
            "example": 4.7
          },
          "isVerified": {
            "type": "boolean",
            "example": true
          },
          "isBlocked": {
            "type": "boolean",
            "example": false
          },
          "isEmailVerified": {
            "type": "boolean",
            "example": true
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "verified",
              "blocked"
            ],
            "example": "verified"
          },
          "lateExcuseCount": {
            "type": "integer",
            "example": 1
          },
          "consecutiveGoodEvents": {
            "type": "integer",
            "example": 3
          },
          "languages": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "arabic",
                "english",
                "french",
                "spanish",
                "german",
                "italian"
              ]
            },
            "example": [
              "arabic",
              "english"
            ]
          },
          "eventCategories": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "wedding",
                "corporate",
                "club",
                "festival",
                "private_party",
                "conference",
                "exhibition",
                "sport_event"
              ]
            },
            "example": [
              "wedding",
              "corporate"
            ]
          },
          "portfolioPicture": {
            "$ref": "#/components/schemas/CloudinaryFile"
          },
          "portfolio": {
            "type": "array",
            "items": {}
          },
          "organizationInfo": {
            "type": "object",
            "nullable": true,
            "description": "Organizer-specific data (description, website).",
            "properties": {
              "description": {
                "type": "string"
              },
              "website": {
                "type": "string",
                "format": "uri"
              }
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}